[Oztechan/CCC#1935] Move premium viewmodel to android
