+++ android/ui/mobile/android-ui-mobile.gradle.kts
@@ -88,13 +88,16 @@ dependencies {
         implementation(project(analytics))
     }
 
+    Modules.Android.ViewModel.apply {
+        implementation(project(premium))
+    }
+
     Modules.Client.ViewModel.apply {
         implementation(project(main))
         implementation(project(calculator))
         implementation(project(currencies))
         implementation(project(settings))
         implementation(project(selectCurrency))
-        implementation(project(premium))
         implementation(project(watchers))
     }
