+++ android/ui/mobile/src/main/kotlin/com/oztechan/ccc/android/ui/mobile/content/premium/PremiumBottomSheet.kt
@@ -20,11 +20,11 @@ import com.oztechan.ccc.android.ui.mobile.util.showSnack
 import com.oztechan.ccc.android.ui.mobile.util.toOldPurchaseList
 import com.oztechan.ccc.android.ui.mobile.util.toPremiumDataList
 import com.oztechan.ccc.android.ui.mobile.util.visibleIf
+import com.oztechan.ccc.android.viewmodel.premium.PremiumEffect
+import com.oztechan.ccc.android.viewmodel.premium.PremiumViewModel
 import com.oztechan.ccc.client.core.analytics.AnalyticsManager
 import com.oztechan.ccc.client.core.analytics.model.ScreenName
 import com.oztechan.ccc.client.core.shared.model.PremiumType
-import com.oztechan.ccc.client.viewmodel.premium.PremiumEffect
-import com.oztechan.ccc.client.viewmodel.premium.PremiumViewModel
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import org.koin.android.ext.android.inject
