+++ buildSrc/src/main/kotlin/Modules.kt
@@ -16,6 +16,7 @@ object Modules {
         }
 
         object ViewModel {
+            const val premium = ":android:viewmodel:premium"
             const val widget = ":android:viewmodel:widget"
         }
     }
@@ -96,7 +97,6 @@ object Modules {
             const val currencies = ":client:viewmodel:currencies"
             const val settings = ":client:viewmodel:settings"
             const val selectCurrency = ":client:viewmodel:selectcurrency"
-            const val premium = ":client:viewmodel:premium"
             const val watchers = ":client:viewmodel:watchers"
         }
     }
