--- client/viewmodel/premium/src/commonMain/kotlin/com/oztechan/ccc/client/viewmodel/premium/di/ClientViewModelPremiumModule.kt
@@ -1,9 +1,0 @@
-package com.oztechan.ccc.client.viewmodel.premium.di
-
-import com.oztechan.ccc.client.core.viewmodel.di.viewModelDefinition
-import com.oztechan.ccc.client.viewmodel.premium.PremiumViewModel
-import org.koin.dsl.module
-
-val clientViewModelPremiumModule = module {
-    viewModelDefinition { PremiumViewModel(get()) }
-}
